fix cases
