+++ benchmark/alpa/benchmark_one_case_gpt_bert_inference.py
@@ -11,7 +11,7 @@ from alpa.util import print_used_time
 
 from util import compute_gpt_parameter_count, compute_gpt_tflops
 from benchmark_parallel_utils import (
-    get_pipeshard_parallel_method, dump_chrome_tracing,
+    get_pipeshard_parallel_method,
     compile_and_benchmark_pipeshard_inference_executable)
 
 
@@ -169,9 +169,7 @@ def benchmark_gpt_inference_internal(model_type,
          profile_driver_time=profile_driver_time)
 
     if profile_stage_execution_time:
-        exec_info = executable.get_stage_execution_info()
-        dump_chrome_tracing(
-            zip(*exec_info),
+        executable.dump_stage_execution_trace(
             f"./chrome_trace/bs={benchmark_case.batch_size},pp={num_manual_pipeline_stages}.json"
         )
