+++ benchmark/alpa/benchmark_parallel_utils.py
@@ -418,67 +418,3 @@ def compile_and_benchmark_pipeshard_inference_executable(
     max_mem_allocated = executable.mesh_group.get_max_memory_allocated()
 
     return latencies, max_mem_allocated, compilation_times, executable
-
-
-def dump_chrome_tracing(timelines: List[tuple], dumpfile: str):
-
-    def get_color(i):
-        color_list = [
-            "thread_state_uninterruptible",
-            "thread_state_iowait",
-            "thread_state_running",
-            "thread_state_runnable",
-            "thread_state_unknown",
-            "background_memory_dump",
-            "light_memory_dump",
-            "detailed_memory_dump",
-            "vsync_highlight_color",
-            "generic_work",
-            "good",
-            "bad",
-            "terrible",
-            "yellow",
-            "olive",
-            "rail_response",
-            "rail_animation",
-            "rail_idle",
-            "rail_load",
-            "startup",
-            "heap_dump_stack_frame",
-            "heap_dump_object_type",
-            "heap_dump_child_node_arrow",
-            "cq_build_running",
-            "cq_build_passed",
-            "cq_build_failed",
-            "cq_build_attempt_runnig",
-            "cq_build_attempt_passed",
-            "cq_build_attempt_failed",
-        ]
-        return color_list[i % len(color_list)]
-
-    slot_list = []
-    for request_id, request_timeline in enumerate(timelines):
-        sorted_timeline = sorted(request_timeline, key=lambda x: x[0])
-
-        for stage_num, (s, e, node_ids, devices) in enumerate(sorted_timeline):
-            for node_id, devices_per_node in zip(node_ids, devices):
-                for device_id in devices_per_node:
-                    slot = {
-                        "name": f"r{request_id}s{stage_num}",
-                        "cat": f"request {request_id}, stage {stage_num}",
-                        "ph": "X",
-                        "pid": node_id,
-                        "tid": device_id,
-                        "ts": float(s) * 1e6,
-                        "dur": float(e - s) * 1e6,
-                        "cname": get_color(request_id)
-                    }
-                    slot_list.append(slot)
-
-    os.makedirs(os.path.dirname(dumpfile), exist_ok=True)
-    with open(dumpfile, "w") as fout:
-        fout.write(
-            json.dumps({
-                "traceEvents": slot_list,
-                "displayTimeUnit": "ms",
-            }))
